[PopularMoviesApp]: finish DetailView
Movie:
We save backdrop url in movie to use it in DetailView as a background.
Also we made Movie class Parcelable to pass it in Intent as object.
MoviesPresenterImpl:
We request and save  backdrop url and average vote of a film.
We changed getImgUrlFor method to build a url for various img sizes.
Also we added Utils class.